requirements.txt
aspose-words==23.5.0
docx2pdf==0.1.8
PySimpleGUI==4.60.5
tk==0.1.0
tqdm==4.65.0
